+++ libs/platform/src/lib/table/components/table-column/table-column.ts
@@ -60,6 +60,9 @@ export abstract class TableColumn {
     /** Whether to apply fd-table-text (text-shadow) to the cell content, if disabled noWrap has no effect. */
     abstract applyText: boolean;
 
+    /** Stores information for the header cell if the ellipsis are visible after the column resize */
+    abstract headerOverflows: boolean;
+
     /** Whether this column is visible */
     abstract responsiveState: FdpColumnResponsiveState;
 }